SALTO-2665: move compact to template and createTemplateExpression
