+++ packages/zendesk-adapter/test/filters/dynamic_content_references.test.ts
@@ -104,7 +104,7 @@ describe('dynamic content references filter', () => {
       expect(instance.value.raw_value).toEqual(new TemplateExpression({
         parts: ['{{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
-          '}}', ' ', '{{notExistsPlaceholder}}', ' ', '{{',
+          '}} {{notExistsPlaceholder}} {{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
           '}}'],
       }))
@@ -112,9 +112,9 @@ describe('dynamic content references filter', () => {
       expect(secondInstance.value.raw_value).toEqual(new TemplateExpression({
         parts: ['{{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
-          '}}', ' ', '{{notExistsPlaceholder}}', ' ', 'bb', '{{',
+          '}} {{notExistsPlaceholder}} bb{{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
-          '}}', 'cc'],
+          '}}cc'],
       }))
     })
   })
@@ -147,7 +147,7 @@ describe('dynamic content references filter', () => {
       expect(instanceCopy.value.raw_value).toEqual(new TemplateExpression({
         parts: ['{{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
-          '}}', ' ', '{{notExistsPlaceholder}}', ' ', '{{',
+          '}} {{notExistsPlaceholder}} {{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
           '}}'],
       }))
@@ -162,9 +162,9 @@ describe('dynamic content references filter', () => {
       expect(secondInstanceCopy.value.raw_value).toEqual(new TemplateExpression({
         parts: ['{{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
-          '}}', ' ', '{{notExistsPlaceholder}}', ' ', 'bb', '{{',
+          '}} {{notExistsPlaceholder}} bb{{',
           new ReferenceExpression(dynamicContentInstance.elemID, dynamicContentInstance),
-          '}}', 'cc'],
+          '}}cc'],
       }))
     })
   })
